% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_pos_tagging.R
\name{rdr_available_models}
\alias{rdr_available_models}
\title{Lists all installed Ripple Down Rules-based Part-Of-Speech Tagger language models}
\usage{
rdr_available_models()
}
\value{
a list of data.frames with elements MORPH, POS and UniversalPOS containing the
language, the location of the dictionary and the location of the ripple down rules
}
\description{
Lists all language models currently installed in the RDRPOSTagger package folder.
These current consists of language models for
\itemize{
 \item{MORPH: }{detailed morphological tagging for languages}
 \item{POS: }{basic parts of speech tagging for languages}
 \item{UniversalPOS: }{universal POS tagging for languages. Based on data from http://universaldependencies.org version 2.0}
}
}
\examples{
models <- rdr_available_models()
models

models$MORPH$language
models$POS$language
models$UniversalPOS$language
}
\seealso{
\code{\link{rdr_model}}
}
