% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{RDRPOSTagger}
\alias{RDRPOSTagger}
\alias{RDRPOSTagger-package}
\title{Parts of Speech Tagging based on the Ripple Down Rules-based Part-Of-Speech Tagger}
\description{
Parts of Speech tagging and morphological tagging based on the
Ripple Down Rules-based Part-Of-Speech Tagger (RDRPOS) available at https://github.com/datquocnguyen/RDRPOSTagger.
Tagging models include Bulgarian, Czech, Dutch, English, French, German, Hindi, Italian, Portuguese, Spanish, Swedish, Thai and Vietnamese.
RDRPOSTagger also supports the pre-trained Universal POS tagging models for 40 languages.
}
\seealso{
\code{\link{rdr_pos}}, \code{\link{rdr_model}}
}
