% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_pos_tagging.R
\name{rdr_model}
\alias{rdr_model}
\title{Set up a Ripple Down Rules-based Part-Of-Speech Tagger for tagging sentences}
\usage{
rdr_model(language, annotation = c("MORPH", "POS", "UniversalPOS"))
}
\arguments{
\item{language}{the language which is one of the languages for the annotation shown in \code{\link{rdr_available_models}}}

\item{annotation}{the type of annotation. Either one of 'MORPH', "POS' or 'UniversalPOS'}
}
\value{
An object of class RDRPOSTagger which is a list with elements model (the location of the dictionary and the rules of that language),
the type of annotation and java objects tagger, initialtagger, dictionary and utility.
This model object can be used to tag sentences based on the specified POS tags.
}
\description{
Set up a Ripple Down Rules-based Part-Of-Speech Tagger for tagging sentences
Possible languages are:
\itemize{
 \item{MORPH: }{Bulgarian, Czech, Dutch, French, German, Portuguese, Spanish, Swedish}
 \item{POS: }{English, French, German, Hindi, Italian, Thai, Vietnamese}
 \item{UniversalPOS: }{Ancient_Greek, Ancient_Greek-PROIEL, Arabic, Basque, Belarusian, 
 Bulgarian, Catalan, Chinese, Coptic, Croatian, Czech, Czech-CAC, Czech-CLTT, 
 Danish, Dutch, Dutch-LassySmall, English, English-LinES, English-ParTUT, Estonian, 
 Finnish, Finnish-FTB, French, French-ParTUT, French-Sequoia, Galician, Galician-TreeGal, 
 German, Gothic, Greek, Hebrew, Hindi, Hungarian, Indonesian, Irish, Italian, 
 Italian-ParTUT, Japanese, Korean, Latin, Latin-ITTB, Latin-PROIEL, Latvian, Lithuanian, 
 Norwegian-Bokmaal, Norwegian-Nynorsk, Old_Church_Slavonic, Persian, Polish, Portuguese, 
 Portuguese-BR, Romanian, Russian, Russian-SynTagRus, Slovak, Slovenian, Slovenian-SST, 
 Spanish, Spanish-AnCora, Swedish, Swedish-LinES, Tamil, Turkish, Urdu, Vietnamese}
}
}
\examples{
\dontrun{
## MORPH models
tagger <- rdr_model(language = "Bulgarian", annotation = "MORPH")
tagger <- rdr_model(language = "Czech", annotation = "MORPH")
tagger <- rdr_model(language = "Dutch", annotation = "MORPH")
tagger <- rdr_model(language = "French", annotation = "MORPH")
tagger <- rdr_model(language = "German", annotation = "MORPH")
tagger <- rdr_model(language = "Portuguese", annotation = "MORPH")
tagger <- rdr_model(language = "Spanish", annotation = "MORPH")
tagger <- rdr_model(language = "Swedish", annotation = "MORPH")
## POS models
tagger <- rdr_model(language = "English", annotation = "POS")
tagger <- rdr_model(language = "French", annotation = "POS")
tagger <- rdr_model(language = "German", annotation = "POS")
tagger <- rdr_model(language = "Hindi", annotation = "POS")
tagger <- rdr_model(language = "Italian", annotation = "POS")
tagger <- rdr_model(language = "Thai", annotation = "POS")
tagger <- rdr_model(language = "Vietnamese", annotation = "POS")
## UniversalPOS models
tagger <- rdr_model(language = "Ancient_Greek", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Ancient_Greek-PROIEL", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Arabic", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Basque", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Belarusian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Bulgarian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Catalan", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Chinese", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Coptic", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Croatian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Czech", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Czech-CAC", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Czech-CLTT", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Danish", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Dutch", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Dutch-LassySmall", annotation = "UniversalPOS")
tagger <- rdr_model(language = "English", annotation = "UniversalPOS")
tagger <- rdr_model(language = "English-LinES", annotation = "UniversalPOS")
tagger <- rdr_model(language = "English-ParTUT", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Estonian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Finnish", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Finnish-FTB", annotation = "UniversalPOS")
tagger <- rdr_model(language = "French", annotation = "UniversalPOS")
tagger <- rdr_model(language = "French-ParTUT", annotation = "UniversalPOS")
tagger <- rdr_model(language = "French-Sequoia", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Galician", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Galician-TreeGal", annotation = "UniversalPOS")
tagger <- rdr_model(language = "German", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Gothic", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Greek", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Hebrew", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Hindi", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Hungarian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Indonesian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Irish", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Italian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Italian-ParTUT", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Japanese", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Korean", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Latin", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Latin-ITTB", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Latin-PROIEL", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Latvian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Lithuanian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Norwegian-Bokmaal", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Norwegian-Nynorsk", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Old_Church_Slavonic", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Persian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Polish", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Portuguese", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Portuguese-BR", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Romanian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Russian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Russian-SynTagRus", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Slovak", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Slovenian", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Slovenian-SST", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Spanish", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Spanish-AnCora", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Swedish", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Swedish-LinES", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Tamil", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Turkish", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Urdu", annotation = "UniversalPOS")
tagger <- rdr_model(language = "Vietnamese", annotation = "UniversalPOS")
}
}
\seealso{
\code{\link{rdr_model}}
}
