% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_pos_tagging.R
\name{rdr_pos}
\alias{rdr_pos}
\title{Part-Of-Speech Tagging for tagging sentences based on Ripple Down Rules}
\usage{
rdr_pos(object, x, sentence.id = seq_along(x))
}
\arguments{
\item{object}{And object of class RDRPOSTagger as returned by \code{\link{rdr_model}}}

\item{x}{a character vector in UTF-8 encoding where each element of the character vector contains only 1 sentence which you like to tag.}

\item{sentence.id}{an identifier of a sentence with the same length as \code{x}}
}
\value{
a data frame with fields sentence.id, word.id, word, word.type where the word.type is the Parts of Speech tag
}
\description{
Part-Of-Speech Tagging for tagging sentences based on Ripple Down Rules
}
\examples{
x <- c("Dus godvermehoeren met pus in alle puisten, zei die schele van Van Bukburg.", 
  "Er was toen dat liedje van tietenkonttieten kont tieten kontkontkont",
  "  ", "", NA)
tagger <- rdr_model(language = "Dutch", annotation = "MORPH")
rdr_pos(tagger, x = x)

tagger <- rdr_model(language = "UD_Dutch", annotation = "UniversalPOS")
rdr_pos(tagger, x = x)

x <- c("Oleg Borisovich Kulik is a Ukrainian-born Russian performance artist, 
  sculptor, photographer and curator.")
tagger <- rdr_model(language = "English", annotation = "POS")
rdr_pos(tagger, x = x)
}
\seealso{
\code{\link{rdr_model}}
}
